split off markdownToProseMirror plugin
